+++ sv-parser-pp/testcases/test11.sv
@@ -1,0 +1,6 @@
+module a;
+initial begin
+    if (`__LINE__ == 0) begin
+    end
+end
+endmodule